refactor/safe_app: move ffi test_utils functions to an ffi module
