+++ safe_app/src/test_utils.rs
@@ -17,15 +17,12 @@
 
 use super::{App, AppContext};
 use super::errors::AppError;
-use ffi_utils::{FFI_RESULT_OK, FfiResult, catch_unwind_cb, from_c_str};
 use futures::{Future, IntoFuture};
 use safe_authenticator::test_utils as authenticator;
 use safe_core::{Client, FutureExt, utils};
-use safe_core::ffi::ipc::req::ContainerPermissions as FfiContainerPermissions;
 use safe_core::ipc::AppExchangeInfo;
-use safe_core::ipc::req::{AuthReq, ContainerPermissions, containers_from_repr_c};
+use safe_core::ipc::req::{AuthReq as NativeAuthReq, ContainerPermissions};
 use std::collections::HashMap;
-use std::os::raw::{c_char, c_void};
 use std::sync::mpsc;
 
 /// Generates an `AppExchangeInfo` structure for a mock application.
@@ -81,81 +78,52 @@ where
     unwrap!(unwrap!(rx.recv()))
 }
 
-// Create registered app with optional app id and access info.
-fn create_app_with_id_or_access(
+/// Create a random app
+pub fn create_app() -> App {
+    create_app_by_req(&create_random_auth_req())
+}
+
+/// Create a random app given an app authorisation request
+pub fn create_app_by_req(auth_req: &NativeAuthReq) -> App {
+    let auth = authenticator::create_account_and_login();
+    let auth_granted = unwrap!(authenticator::register_app(&auth, auth_req));
+    unwrap!(App::registered(
+        auth_req.app.id.clone(),
+        auth_granted,
+        || (),
+    ))
+}
+
+/// Create an app authorisation request with optional app id and access info.
+pub fn create_auth_req(
     app_id: Option<String>,
     access_info: Option<HashMap<String, ContainerPermissions>>,
-) -> App {
-    let auth = authenticator::create_account_and_login();
-
+) -> NativeAuthReq {
     let mut app_info = gen_app_exchange_info();
     if let Some(app_id) = app_id {
         app_info.id = app_id;
     }
-    let app_id = app_info.id.clone();
 
     let (app_container, containers) = match access_info {
         Some(access_info) => (true, access_info),
         None => (false, HashMap::new()),
     };
 
-    let auth_granted = unwrap!(authenticator::register_app(
-        &auth,
-        &AuthReq {
-            app: app_info,
-            app_container,
-            containers,
-        },
-    ));
-
-    unwrap!(App::registered(app_id, auth_granted, || ()))
+    NativeAuthReq {
+        app: app_info,
+        app_container,
+        containers,
+    }
 }
 
 /// Create registered app with a random id.
-pub fn create_app() -> App {
-    create_app_with_id_or_access(None, None)
+pub fn create_random_auth_req() -> NativeAuthReq {
+    create_auth_req(None, None)
 }
 
 /// Create registered app with a random id and grant it access to the specified containers.
-pub fn create_app_with_access(access_info: HashMap<String, ContainerPermissions>) -> App {
-    create_app_with_id_or_access(None, Some(access_info))
-}
-
-/// Creates a random app instance for testing.
-#[no_mangle]
-#[allow(unsafe_code)]
-pub unsafe extern "C" fn test_create_app(
-    app_id: *const c_char,
-    user_data: *mut c_void,
-    o_cb: extern "C" fn(user_data: *mut c_void,
-                        result: *const FfiResult,
-                        app: *mut App),
-) {
-    catch_unwind_cb(user_data, o_cb, || -> Result<(), AppError> {
-        let app_id = from_c_str(app_id)?;
-        let app = create_app_with_id_or_access(Some(app_id), None);
-        o_cb(user_data, FFI_RESULT_OK, Box::into_raw(Box::new(app)));
-        Ok(())
-    })
-}
-
-/// Create a random app instance for testing, with access to containers.
-#[no_mangle]
-#[allow(unsafe_code)]
-pub unsafe extern "C" fn test_create_app_with_access(
-    app_id: *const c_char,
-    access_info: *const FfiContainerPermissions,
-    access_info_len: usize,
-    user_data: *mut c_void,
-    o_cb: extern "C" fn(user_data: *mut c_void,
-                        result: *const FfiResult,
-                        o_app: *mut App),
-) {
-    catch_unwind_cb(user_data, o_cb, || -> Result<(), AppError> {
-        let app_id = from_c_str(app_id)?;
-        let containers = containers_from_repr_c(access_info, access_info_len)?;
-        let app = create_app_with_id_or_access(Some(app_id), Some(containers));
-        o_cb(user_data, FFI_RESULT_OK, Box::into_raw(Box::new(app)));
-        Ok(())
-    })
+pub fn create_auth_req_with_access(
+    access_info: HashMap<String, ContainerPermissions>,
+) -> NativeAuthReq {
+    create_auth_req(None, Some(access_info))
 }
